Refactor. Move common validation annotations to common.web.validator package.
