+++ src/main/java/com/unvise/bankingsystemapp/domain/common/web/validator/FieldUnmatchedValidator.java
@@ -1,37 +1,35 @@
-package com.unvise.bankingsystemapp.domain.auth.web.validator;
+package com.unvise.bankingsystemapp.domain.common.web.validator;
 
-import com.unvise.bankingsystemapp.domain.currency.web.validator.FieldUnmatched;
 import org.springframework.beans.BeanWrapperImpl;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
-public class MatchedFieldValidator implements ConstraintValidator<MatchedField, Object> {
+public class FieldUnmatchedValidator implements ConstraintValidator<FieldUnmatched, Object> {
 
     private String firstFieldName;
     private String secondFieldName;
     private String message;
-
+    
     @Override
-    public void initialize(MatchedField constraintAnnotation) {
+    public void initialize(FieldUnmatched constraintAnnotation) {
         firstFieldName = constraintAnnotation.first();
         secondFieldName = constraintAnnotation.second();
         message = constraintAnnotation.message();
     }
 
     @Override
-    public boolean isValid(Object o, ConstraintValidatorContext context) {
+    public boolean isValid(Object value, ConstraintValidatorContext context) {
         boolean valid = true;
-
         try {
-            BeanWrapperImpl beanWrapper = new BeanWrapperImpl(o);
+            BeanWrapperImpl beanWrapper = new BeanWrapperImpl(value);
             Object firstObj = beanWrapper.getPropertyValue(firstFieldName);
             Object secondObj = beanWrapper.getPropertyValue(secondFieldName);
 
             valid = firstObj == null
                     && secondObj == null
                     || firstObj != null
-                    && firstObj.equals(secondObj);
+                    && !firstObj.equals(secondObj);
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -45,4 +43,5 @@ public class MatchedFieldValidator implements ConstraintValidator<MatchedField, 
 
         return valid;
     }
+
 }
